requirements.txt
altair==4.2.2
atomicwrites==1.4.0
attrs==23.1.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
collection==0.1.6
contourpy==1.0.7
cycler==0.11.0
deta==1.1.0
et-xmlfile==1.1.0
extra-streamlit-components==0.1.56
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gsheetsdb==0.1.13.1
gspread==5.9.0
httplib2==0.22.0
importlib-metadata==6.6.0
inflection==0.5.1
itertable==2.1.0
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
mypy-extensions==0.4.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.5.0
plotly==5.14.1
plotly-express==0.4.1
ply==3.11
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydeck==0.8.1b0
PyJWT==2.7.0
pyls-spyder==0.4.0
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQtWebEngine==5.15.4
pyrsistent==0.19.3
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
pytz==2023.3
pywin32==305.1
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.1
smmap==5.0.0
st-star-rating==0.0.3
statsmodels==0.14.0
streamlit==1.22.0
streamlit-authenticator==0.2.2
tenacity==8.2.2
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.2
trubrics==1.3.7
typeguard==2.13.3
typer==0.9.0
tzdata==2023.3
tzlocal==5.0.1
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
win32-setctime==1.1.0
zipp==3.15.0